# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Express.js starter/boilerplate by @CC (CodeContinent)",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b>"
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Todo",
      "description": "Endpoints for Todo API"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/todo/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "todo-item-title"
        },
        "completed": {
          "type": "boolean",
          "example": false
        },
        "createAt": {
          "type": "string",
          "example": "Thu, 11 May 2023 10:25:00 GMT"
        },
        "updatedAt": {
          "type": "string",
          "example": "Thu, 11 May 2023 10:25:00 GMT"
        }
      }
    },
    "addTodo": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "todo-item-title"
        }
      },
      "required": ["title"]
    }
  }
}